Aula 3 concluída ✅

+++ server/src/routes/memories.ts
@@ -4,8 +4,13 @@ import { prisma } from '../lib/prisma'
 import { z } from 'zod'
 
 export const memoriesRoutes = async (app: FastifyInstance) => {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify()
+  })
+
+  app.get('/memories', async (request) => {
     const memories = await prisma.memory.findMany({
+      where: { userId: request.user.sub },
       orderBy: {
         createAt: 'asc',
       },
@@ -20,7 +25,7 @@ export const memoriesRoutes = async (app: FastifyInstance) => {
     })
   })
 
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -28,6 +33,10 @@ export const memoriesRoutes = async (app: FastifyInstance) => {
     const { id } = paramsSchema.parse(request.params)
 
     const memory = await prisma.memory.delete({ where: { id } })
+
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
     return memory
   })
@@ -46,14 +55,14 @@ export const memoriesRoutes = async (app: FastifyInstance) => {
         content,
         coverUrl,
         isPublic,
-        userId: '760db87e-4297-4a3e-910c-e08c3229d67e',
+        userId: request.user.sub,
       },
     })
 
     return memory
   })
 
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -68,7 +77,15 @@ export const memoriesRoutes = async (app: FastifyInstance) => {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: { id },
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
       where: { id },
       data: {
         content,
@@ -80,12 +97,20 @@ export const memoriesRoutes = async (app: FastifyInstance) => {
     return memory
   })
 
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
 
     const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: { id },
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
     await prisma.memory.delete({ where: { id } })
   })
